Updated testing
