+++ .gitignore
@@ -3,5 +3,6 @@ package/
 *env/
 *.zip
 __pycache__/
+*.pytest_cache
 *.py[cod]
 *$py.class
